update package dependencies
PmWiki requires PHP, but doesn't actually depend on anything else.
A webserver is required, so the package is set to Suggest it, but
even this is not required because there is a "Standalone" cookbook
recipe available:

  http://www.pmwiki.org/wiki/Cookbook/Standalone

Additionally the pmwiki package should not depend on the 'php5'
package, becuase PmWiki does not depend on Apache, and works with
several other webservers:

  http://www.pmwiki.org/wiki/PmWiki/Requirements

Since the PHP binaries seem to actually be shipped with the
php5-common package, that is the only dependency.